SUPDrv: warnings
svn:sync-xref-src-repo-rev: r109238

+++ src/VBox/HostDrivers/Support/SUPLib.cpp
@@ -1,4 +1,4 @@
-/* $Id: SUPLib.cpp 62490 2016-07-22 18:41:49Z knut.osmundsen@oracle.com $ */
+/* $Id: SUPLib.cpp 62665 2016-07-28 23:44:29Z knut.osmundsen@oracle.com $ */
 /** @file
  * VirtualBox Support Library - Common code.
  */
@@ -1985,7 +1985,7 @@ SUPR3DECL(int) SUPR3TracerRegisterModule(uintptr_t hModNative, const char *pszMo
      * Create a string table for the function names in the location array.
      * It's somewhat easier to do that here than from ring-0.
      */
-    size_t const        cProbeLocs  = pVtgHdr->cbProbeLocs
+    uint32_t const      cProbeLocs  = pVtgHdr->cbProbeLocs
                                     / (pVtgHdr->cBits == 32 ? sizeof(VTGPROBELOC32) : sizeof(VTGPROBELOC64));
     PVTGPROBELOC        paProbeLocs = (PVTGPROBELOC)((uintptr_t)pVtgHdr + pVtgHdr->offProbeLocs);
     PSUPDRVTRACERSTRTAB pStrTab     = supr3TracerCreateStrTab((PVTGPROBELOC32)paProbeLocs,
